Cov-class {rrcov}	R Documentation

Class "Cov" – a base class for estimates of multivariate location and scatter

Description

The class Cov represents an estimate of the multivariate location and scatter of a data set. The objects of class Cov contain the classical estimates and serve as base for deriving other estimates, i.e. different types of robust estimates.

Objects from the Class

Objects can be created by calls of the form new("Cov", ...), but the usual way of creating Cov objects is a call to the function Cov which serves as a constructor.

Slots

call:

Object of class "language"

cov:

covariance matrix

center:

location

n.obs:

number of observations used for the computation of the estimates

mah:

mahalanobis distances

det:

determinant

flag:

flags (FALSE if suspected an outlier)

method:

a character string describing the method used to compute the estimate: "Classic"

singularity:

a list with singularity information for the covariance matrix (or NULL of not singular)

X:

data

Methods

getCenter

signature(obj = "Cov"): location vector

getCov

signature(obj = "Cov"): covariance matrix

getCorr

signature(obj = "Cov"): correlation matrix

getData

signature(obj = "Cov"): data frame

getDistance

signature(obj = "Cov"): distances

getEvals

signature(obj = "Cov"): Computes and returns the eigenvalues of the covariance matrix

getDet

signature(obj = "Cov"): Computes and returns the determinant of the covariance matrix (or 0 if the covariance matrix is singular)

getShape

signature(obj = "Cov"): Computes and returns the shape matrix corresponding to the covariance matrix (i.e. the covariance matrix scaled to have determinant =1)

getFlag

signature(obj = "Cov"): Flags observations as outliers if the corresponding mahalanobis distance is larger then qchisq(prob, p) where prob defaults to 0.975.

isClassic

signature(obj = "Cov"): returns TRUE by default. If necessary, the robust classes will override

plot

signature(x = "Cov"): plot the object

show

signature(object = "Cov"): display the object

summary

signature(object = "Cov"): calculate summary information

Author(s)

Valentin Todorov valentin.todorov@chello.at

References

Todorov V & Filzmoser P (2009), An Object Oriented Framework for Robust Multivariate Analysis. Journal of Statistical Software, 32(3), 1–47. doi: 10.18637/jss.v032.i03.

Examples

showClass("Cov")

[Package rrcov version 1.7-7 Index]