CovClassic {rrcov}	R Documentation

Classical Estimates of Multivariate Location and Scatter

Description

Computes the classical estimates of multivariate location and scatter. Returns an S4 class CovClassic with the estimated center, cov, Mahalanobis distances and weights based on these distances.

Usage

    CovClassic(x, unbiased=TRUE)
    Cov(x, unbiased=TRUE)

Arguments

x
a matrix or data frame. As usual, rows are observations and columns are variables.

unbiased
whether to return the unbiased estimate of the covariance matrix. Default is unbiased = TRUE

Value

An object of class "CovClassic".

Author(s)

Valentin Todorov valentin.todorov@chello.at

References

Todorov V & Filzmoser P (2009), An Object Oriented Framework for Robust Multivariate Analysis. Journal of Statistical Software, 32(3), 1–47. doi: 10.18637/jss.v032.i03.

See Also

Cov-class, CovClassic-class

Examples

data(hbk)
hbk.x <- data.matrix(hbk[, 1:3])
cv <- CovClassic(hbk.x)
cv
summary(cv)
plot(cv)

[Package rrcov version 1.7-7 Index]