CovClassic-class {rrcov}	R Documentation

Class "CovClassic" - classical estimates of multivariate location and scatter

Description

The class CovClassic represents an estimate of the multivariate location and scatter of a data set. The objects of class CovClassic contain the classical estimates.

Objects from the Class

Objects can be created by calls of the form new("CovClassic", ...), but the usual way of creating CovClassic objects is a call to the function CovClassic which serves as a constructor.

Slots

call:

Object of class "language"

cov:

covariance matrix

center:

location

n.obs:

number of observations used for the computation of the estimates

mah:

mahalanobis distances

method:

a character string describing the method used to compute the estimate: "Classic"

singularity:

a list with singularity information for the ocvariance matrix (or NULL of not singular)

X:

data

Methods

getCenter

signature(obj = "CovClassic"): location vector

getCov

signature(obj = "CovClassic"): covariance matrix

getCorr

signature(obj = "CovClassic"): correlation matrix

getData

signature(obj = "CovClassic"): data frame

getDistance

signature(obj = "CovClassic"): distances

getEvals

signature(obj = "CovClassic"): Computes and returns the eigenvalues of the covariance matrix

plot

signature(x = "CovClassic"): plot the object

show

signature(object = "CovClassic"): display the object

summary

signature(object = "CovClassic"): calculate summary information

Author(s)

Valentin Todorov valentin.todorov@chello.at

References

Todorov V & Filzmoser P (2009), An Object Oriented Framework for Robust Multivariate Analysis. Journal of Statistical Software, 32(3), 1–47. doi: 10.18637/jss.v032.i03.

Examples

data(hbk)
hbk.x <- data.matrix(hbk[, 1:3])
cv <- CovClassic(hbk.x)
cv
summary(cv)
plot(cv)

[Package rrcov version 1.7-7 Index]