CovControl-class {rrcov}	R Documentation

Class "CovControl" is a VIRTUAL base control class

Description

The class "CovControl" is a VIRTUAL base control class for the derived classes representing the control parameters for the different robust methods

Arguments

trace
whether to print intermediate results. Default is trace = FALSE

tolSolve
numeric tolerance to be used for inversion (solve) of the covariance matrix in mahalanobis.

Objects from the Class

A virtual Class: No objects may be created from it.

Methods

No methods defined with class "CovControl" in the signature.

Author(s)

Valentin Todorov valentin.todorov@chello.at

References

Todorov V & Filzmoser P (2009), An Object Oriented Framework for Robust Multivariate Analysis. Journal of Statistical Software, 32(3), 1–47. doi: 10.18637/jss.v032.i03.

[Package rrcov version 1.7-7 Index]