CovControlMMest {rrcov}	R Documentation

Constructor function for objects of class "CovControlMMest"

Description

This function will create a control object CovControlMMest containing the control parameters for CovMMest

Usage

    CovControlMMest(bdp = 0.5, eff=0.95, maxiter = 50, sest=CovControlSest(),
        trace = FALSE, tolSolve = 1e-7)

Arguments

bdp
a numeric value specifying the required breakdown point. Allowed values are between 0.5 and 1 and the default is 0.5

eff
a numeric value specifying the required efficiency for the MM estimates. Default is eff=0.95.

sest
an CovControlSest object containing control parameters for the initial S-estimate.

maxiter
maximum number of iterations allowed in the computation of the MM-estimate. Defaults to 150.

trace
whether to print intermediate results. Default is trace = FALSE.

tolSolve
numeric tolerance to be used as a convergence tolerance for the MM-iteration.

Value

A CovControlSest object.

Author(s)

Valentin Todorov valentin.todorov@chello.at

References

Todorov V & Filzmoser P (2009), An Object Oriented Framework for Robust Multivariate Analysis. Journal of Statistical Software, 32(3), 1–47. doi: 10.18637/jss.v032.i03.

Examples

   ## the following two statements are equivalent
    ctrl1 <- new("CovControlMMest", bdp=0.25)
    ctrl2 <- CovControlMMest(bdp=0.25)

    data(hbk)
    CovMMest(hbk, control=ctrl1)
    
    
    

[Package rrcov version 1.7-7 Index]