CovControlMMest-class {rrcov}	R Documentation

Class 'CovControlMMest' - contains control parameters for "CovMMest"

Description

This class extends the CovControl class and contains the control parameters for CovMMest

Objects from the Class

Objects can be created by calls of the form new("CovControlMMest", ...) or by calling the constructor-function CovControlMMest.

Slots

bdp

a numeric value specifying the required breakdown point. Allowed values are between 0.5 and 1 and the default is bdp=0.5.

eff

a numeric value specifying the required efficiency for the MM estimates. Default is eff=0.95.

sest

an CovControlSest object containing control parameters for the initial S-estimate.

maxiter

maximum number of iterations allowed in the computation of the MM-estimate. Default is maxiter=50.

trace, tolSolve:

from the "CovControl" class. tolSolve is used as a convergence tolerance for the MM-iteration.

Extends

Class "CovControl", directly.

Methods

restimate

signature(obj = "CovControlMMest"): the generic function restimate allowes the different methods for robust estimation to be used polymorphically - this function will call CovMMest passing it the control object and will return the obtained CovRobust object

Author(s)

Valentin Todorov valentin.todorov@chello.at

References

Todorov V & Filzmoser P (2009), An Object Oriented Framework for Robust Multivariate Analysis. Journal of Statistical Software, 32(3), 1–47. doi: 10.18637/jss.v032.i03.

Examples

    ## the following two statements are equivalent
    ctrl1 <- new("CovControlMMest", bdp=0.25)
    ctrl2 <- CovControlMMest(bdp=0.25)

    data(hbk)
    CovMMest(hbk, control=ctrl1)

[Package rrcov version 1.7-7 Index]