CovControlMcd-class {rrcov}	R Documentation

Class 'CovControlMcd' - contains control parameters for CovMcd

Description

This class extends the CovControl class and contains the control parameters for "CovMcd"

Objects from the Class

Objects can be created by calls of the form new("CovControlMcd", ...) or by calling the constructor-function CovControlMcd.

Slots

alpha:

numeric parameter controlling the size of the subsets over which the determinant is minimized, i.e., alpha*n observations are used for computing the determinant. Allowed values are between 0.5 and 1 and the default is 0.5.

nsamp

number of subsets used for initial estimates or "best", "exact" or "deterministic". Default is nsamp = 500. For nsamp="best" exhaustive enumeration is done, as long as the number of trials does not exceed 5000. For "exact", exhaustive enumeration will be attempted however many samples are needed. In this case a warning message will be displayed saying that the computation can take a very long time.

For "deterministic", the deterministic MCD is computed; as proposed by Hubert et al. (2012) it starts from the h most central observations of six (deterministic) estimators.

scalefn

function to compute a robust scale estimate or character string specifying a rule determining such a function.

maxcsteps

maximal number of concentration steps in the deterministic MCD; should not be reached.

seed:

starting value for random generator. Default is seed = NULL

use.correction:

whether to use finite sample correction factors. Default is use.correction=TRUE.

trace, tolSolve:

from the "CovControl" class.

Extends

Class "CovControl", directly.

Methods

restimate

signature(obj = "CovControlMcd"): the generic function restimate allows the different methods for robust estimation to be used polymorphically - this function will call CovMcd passing it the control object and will return the obtained CovRobust object

Author(s)

Valentin Todorov valentin.todorov@chello.at

References

Todorov V & Filzmoser P (2009), An Object Oriented Framework for Robust Multivariate Analysis. Journal of Statistical Software, 32(3), 1–47. doi: 10.18637/jss.v032.i03.

Examples

    ## the following two statements are equivalent
    ctrl1 <- new("CovControlMcd", alpha=0.75)
    ctrl2 <- CovControlMcd(alpha=0.75)

    data(hbk)
    CovMcd(hbk, control=ctrl1)

[Package rrcov version 1.7-7 Index]