CovControlMest {rrcov}	R Documentation

Constructor function for objects of class "CovControlMest"

Description

This function will create a control object CovControlMest containing the control parameters for CovMest

Usage

CovControlMest(r = 0.45, arp = 0.05, eps = 0.001, maxiter = 120)

Arguments

r
a numeric value specifying the required breakdown point. Allowed values are between (n - p)/(2 * n) and 1 and the default is 0.45

arp
a numeric value specifying the asympthotic rejection point, i.e. the fraction of points receiving zero weight (see Rocke (1996)). Default is 0.05

eps
a numeric value specifying the relative precision of the solution of the M-estimate. Defaults to 1e-3

maxiter
maximum number of iterations allowed in the computation of the M-estimate. Defaults to 120

Value

A CovControlMest object

Author(s)

Valentin Todorov valentin.todorov@chello.at

References

Todorov V & Filzmoser P (2009), An Object Oriented Framework for Robust Multivariate Analysis. Journal of Statistical Software, 32(3), 1–47. doi: 10.18637/jss.v032.i03.

Examples

    ## the following two statements are equivalent
    ctrl1 <- new("CovControlMest", r=0.4)
    ctrl2 <- CovControlMest(r=0.4)

    data(hbk)
    CovMest(hbk, control=ctrl1)

[Package rrcov version 1.7-7 Index]