CovControlMrcd {rrcov}	R Documentation

Constructor function for objects of class "CovControlMrcd"

Description

This function will create a control object CovControlMrcd containing the control parameters for CovMrcd

Usage

CovControlMrcd(alpha = 0.5, h=NULL, maxcsteps=200, rho=NULL, 
    target=c("identity", "equicorrelation"), maxcond=50,
    trace= FALSE)

Arguments

alpha
numeric parameter controlling the size of the subsets over which the determinant is minimized, i.e., alpha*n observations are used for computing the determinant. Allowed values are between 0.5 and 1 and the default is 0.5.

h
the size of the subset (can be between ceiling(n/2) and n). Normally NULL and then it h will be calculated as h=ceiling(alpha*n). If h is provided, alpha will be calculated as alpha=h/n.

maxcsteps
maximal number of concentration steps in the deterministic MCD; should not be reached.

rho
regularization parameter. Normally NULL and will be estimated from the data.

target
structure of the robust positive definite target matrix: a) "identity": target matrix is diagonal matrix with robustly estimated univariate scales on the diagonal or b) "equicorrelation": non-diagonal target matrix that incorporates an equicorrelation structure (see (17) in paper). Default is target="identity"

maxcond
maximum condition number allowed (see step 3.4 in algorithm 1). Default is maxcond=50

trace
whether to print intermediate results. Default is trace = FALSE

Value

A CovControlMrcd object

Author(s)

Valentin Todorov valentin.todorov@chello.at

References

Todorov V & Filzmoser P (2009), An Object Oriented Framework for Robust Multivariate Analysis. Journal of Statistical Software, 32(3), 1–47. doi: 10.18637/jss.v032.i03.

Examples

    ## the following two statements are equivalent
    ctrl1 <- new("CovControlMrcd", alpha=0.75)
    ctrl2 <- CovControlMrcd(alpha=0.75)

    data(hbk)
    CovMrcd(hbk, control=ctrl1)

[Package rrcov version 1.7-7 Index]