CovControlSde {rrcov}	R Documentation

Constructor function for objects of class "CovControlSde"

Description

This function will create a control object CovControlSde containing the control parameters for CovSde

Usage

CovControlSde(nsamp = 0, maxres = 0, tune = 0.95, eps = 0.5, prob = 0.99,
    seed = NULL, trace = FALSE, tolSolve = 1e-14)

Arguments

nsamp
a positive integer giving the number of resamples required; nsamp may not be reached if too many of the p-subsamples, chosen out of the observed vectors, are in a hyperplane. If nsamp = 0 all possible subsamples are taken. If nsamp is omitted, it is calculated to provide a breakdown point of eps with probability prob.

maxres
a positive integer specifying the maximum number of resamples to be performed including those that are discarded due to linearly dependent subsamples. If maxres is omitted it will be set to 2 times nsamp.

tune
a numeric value between 0 and 1 giving the fraction of the data to receive non-zero weight. Defaults to 0.95.

prob
a numeric value between 0 and 1 specifying the probability of high breakdown point; used to compute nsamp when nsamp is omitted. Defaults to 0.99.

eps
a numeric value between 0 and 0.5 specifying the breakdown point; used to compute nsamp when nresamp is omitted. Defaults to 0.5.

seed
starting value for random generator. Default is seed = NULL.

trace
whether to print intermediate results. Default is trace = FALSE.

tolSolve
numeric tolerance to be used for inversion (solve) of the covariance matrix in mahalanobis.

Value

A CovControlSde object.

Author(s)

Valentin Todorov valentin.todorov@chello.at

References

Todorov V & Filzmoser P (2009), An Object Oriented Framework for Robust Multivariate Analysis. Journal of Statistical Software, 32(3), 1–47. doi: 10.18637/jss.v032.i03.

Examples

    ## the following two statements are equivalent
    ctrl1 <- new("CovControlSde", nsamp=2000)
    ctrl2 <- CovControlSde(nsamp=2000)

    data(hbk)
    CovSde(hbk, control=ctrl1)

[Package rrcov version 1.7-7 Index]