CovControlSest {rrcov}	R Documentation

Constructor function for objects of class "CovControlSest"

Description

This function will create a control object CovControlSest containing the control parameters for CovSest

Usage

    CovControlSest(bdp = 0.5, arp = 0.1, eps = 1e-5, maxiter = 120,
        nsamp = 500, seed = NULL, trace = FALSE, tolSolve = 1e-14, method= "sfast")

Arguments

bdp
a numeric value specifying the required breakdown point. Allowed values are between (n - p)/(2 * n) and 1 and the default is 0.45

arp
a numeric value specifying the asympthotic rejection point (for the Rocke type S estimates), i.e. the fraction of points receiving zero weight (see Rocke (1996)). Default is 0.1

eps
a numeric value specifying the relative precision of the solution of the S-estimate (bisquare and Rocke type). Defaults to 1e-5.

maxiter
maximum number of iterations allowed in the computation of the S-estimate (bisquare and Rocke type). Defaults to 120.

nsamp
the number of random subsets considered. Default is nsamp = 500.

seed
starting value for random generator. Default is seed = NULL.

trace
whether to print intermediate results. Default is trace = FALSE.

tolSolve
numeric tolerance to be used for inversion (solve) of the covariance matrix in mahalanobis.

method
Which algorithm to use: 'sfast'=FAST-S or 'surreal'=SURREAL

Value

A CovControlSest object.

Author(s)

Valentin Todorov valentin.todorov@chello.at

References

Todorov V & Filzmoser P (2009), An Object Oriented Framework for Robust Multivariate Analysis. Journal of Statistical Software, 32(3), 1–47. doi: 10.18637/jss.v032.i03.

Examples

    ## the following two statements are equivalent
    ctrl1 <- new("CovControlSest", bdp=0.4)
    ctrl2 <- CovControlSest(bdp=0.4)

    data(hbk)
    CovSest(hbk, control=ctrl1)

[Package rrcov version 1.7-7 Index]