CovMest-class {rrcov}	R Documentation

Constrained M-estimates of Multivariate Location and Scatter

Description

This class, derived from the virtual class "CovRobust" accomodates constrained M-Estimates of multivariate location and scatter based on the translated biweight function (‘t-biweight’) using a High breakdown point initial estimate (Minimum Covariance Determinant - ‘Fast MCD’)

Objects from the Class

Objects can be created by calls of the form new("CovMest", ...), but the usual way of creating CovMest objects is a call to the function CovMest which serves as a constructor.

Slots

vt:

Object of class "vector" - vector of weights (v)

iter, crit, wt:

from the "CovRobust" class.

call, cov, center, n.obs, mah, method, singularity, X:

from the "Cov" class.

Extends

Class "CovRobust", directly. Class "Cov", by class "CovRobust".

Methods

No methods defined with class "CovMest" in the signature.

Author(s)

Valentin Todorov valentin.todorov@chello.at

References

Todorov V & Filzmoser P (2009), An Object Oriented Framework for Robust Multivariate Analysis. Journal of Statistical Software, 32(3), 1–47. doi: 10.18637/jss.v032.i03.

See Also

CovMest, Cov-class, CovRobust-class

Examples

showClass("CovMest")

[Package rrcov version 1.7-7 Index]