CovMrcd-class {rrcov}	R Documentation

MRCD Estimates of Multivariate Location and Scatter

Description

This class, derived from the virtual class "CovRobust" accomodates MRCD Estimates of multivariate location and scatter computed by a variant of the ‘Fast MCD’ algorithm.

Objects from the Class

Objects can be created by calls of the form new("CovMrcd", ...), but the usual way of creating CovMrcd objects is a call to the function CovMrcd which serves as a constructor.

Slots

alpha:

Object of class "numeric" - the size of the subsets over which the determinant is minimized (the default is (n+p+1)/2)

quan:

Object of class "numeric" - the number of observations on which the MCD is based. If quan equals n.obs, the MCD is the classical covariance matrix.

best:

Object of class "Uvector" - the best subset found and used for computing the raw estimates. The size of best is equal to quan

cnp2:

Object of class "numeric" - containing the consistency correction factor of the estimate of the covariance matrix.

icov:

The inverse of the covariance matrix.

rho:

The estimated regularization parameter.

target:

The estimated target matrix.

crit:

from the "CovRobust" class.

call, cov, center, n.obs, mah, method, X:

from the "Cov" class.

Extends

Class "CovRobust", directly. Class "Cov", by class "CovRobust".

Methods

No methods defined with class "CovMrcd" in the signature.

Author(s)

Valentin Todorov valentin.todorov@chello.at

References

Todorov V & Filzmoser P (2009), An Object Oriented Framework for Robust Multivariate Analysis. Journal of Statistical Software, 32(3), 1–47. doi: 10.18637/jss.v032.i03.

See Also

CovMrcd, Cov-class, CovRobust-class, CovMcd-class

Examples

showClass("CovMrcd")

[Package rrcov version 1.7-7 Index]