LdaPP-class {rrcov}	R Documentation

Class "LdaPP" - Robust method for Linear Discriminant Analysis by Projection-pursuit

Description

The class LdaPP represents an algorithm for robust linear discriminant analysis by projection-pursuit approach. The objects of class LdaPP contain the results of the robust linear discriminant analysis by projection-pursuit approach.

Objects from the Class

Objects can be created by calls of the form new("LdaPP", ...) but the usual way of creating LdaPP objects is a call to the function LdaPP which serves as a constructor.

Slots

call:

The (matched) function call.

prior:

Prior probabilities used, default to group proportions

counts:

number of observations in each class

center:

the group means

cov:

the common covariance matrix

raw.ldf:

a matrix containing the raw linear discriminant functions - see Details in LdaPP

raw.ldfconst:

a vector containing the raw constants of each raw linear discriminant function - see Details in LdaPP

ldf:

a matrix containing the linear discriminant functions

ldfconst:

a vector containing the constants of each linear discriminant function

method:

a character string giving the estimation method used

X:

the training data set (same as the input parameter x of the constructor function)

grp:

grouping variable: a factor specifying the class for each observation.

Extends

Class "LdaRobust", directly. Class "Lda", by class "LdaRobust", distance 2.

Methods

predict

signature(object = "LdaPP"): calculates prediction using the results in object. An optional data frame or matrix in which to look for variables with which to predict. If omitted, the training data set is used. If the original fit used a formula or a data frame or a matrix with column names, newdata must contain columns with the same names. Otherwise it must contain the same number of columns, to be used in the same order. If the argument raw=TRUE is set the raw (obtained by the first approximation algorithm) linear discriminant function and constant will be used.

Author(s)

Valentin Todorov valentin.todorov@chello.at and Ana Pires apires@math.ist.utl.pt

References

Pires, A. M. and A. Branco, J. (2010) Projection-pursuit approach to robust linear discriminant analysis Journal Multivariate Analysis, Academic Press, Inc., 101, 2464–2485.

Todorov V & Filzmoser P (2009), An Object Oriented Framework for Robust Multivariate Analysis. Journal of Statistical Software, 32(3), 1–47. doi: 10.18637/jss.v032.i03.

See Also

LdaRobust-class, Lda-class, LdaClassic, LdaClassic-class, Linda, Linda-class

Examples

showClass("LdaPP")

[Package rrcov version 1.7-7 Index]