LdaRobust-class {rrcov}	R Documentation

Class "LdaRobust" is a virtual base class for all robust LDA classes

Description

The class LdaRobust searves as a base class for deriving all other classes representing the results of the robust Linear Discriminant Analysis methods

Objects from the Class

A virtual Class: No objects may be created from it.

Slots

call:

The (matched) function call.

prior:

Prior probabilities used, default to group proportions

counts:

number of observations in each class

center:

the group means

cov:

the common covariance matrix

ldf:

a matrix containing the linear discriminant functions

ldfconst:

a vector containing the constants of each linear discriminant function

method:

a character string giving the estimation method used

X:

the training data set (same as the input parameter x of the constructor function)

grp:

grouping variable: a factor specifying the class for each observation.

Extends

Class "Lda", directly.

Methods

No methods defined with class "LdaRobust" in the signature.

Author(s)

Valentin Todorov valentin.todorov@chello.at

References

Todorov V & Filzmoser P (2009), An Object Oriented Framework for Robust Multivariate Analysis. Journal of Statistical Software, 32(3), 1–47. doi: 10.18637/jss.v032.i03.

See Also

Lda-class, LdaClassic-class,

Examples

showClass("LdaRobust")

[Package rrcov version 1.7-7 Index]