PcaCov-class {rrcov}	R Documentation

Class "PcaCov" - Robust PCA based on a robust covariance matrix

Description

Robust PCA are obtained by replacing the classical covariance matrix by a robust covariance estimator. This can be one of the available in rrcov estimators, i.e. MCD, OGK, M, S or Stahel-Donoho estimator.

Objects from the Class

Objects can be created by calls of the form new("PcaCov", ...) but the usual way of creating PcaCov objects is a call to the function PcaCov which serves as a constructor.

Slots

quan:

Object of class "numeric" The quantile h used throughout the algorithm

call, center, rank, loadings, eigenvalues, scores, k, sd, od, cutoff.sd, cutoff.od, flag, n.obs, eig0, totvar0:

from the "Pca" class.

Extends

Class "PcaRobust", directly. Class "Pca", by class "PcaRobust", distance 2.

Methods

getQuan

signature(obj = "PcaCov"): ...

Author(s)

Valentin Todorov valentin.todorov@chello.at

References

Todorov V & Filzmoser P (2009), An Object Oriented Framework for Robust Multivariate Analysis. Journal of Statistical Software, 32(3), 1–47. doi: 10.18637/jss.v032.i03.

See Also

PcaRobust-class, Pca-class, PcaClassic, PcaClassic-class

Examples

showClass("PcaCov")

[Package rrcov version 1.7-7 Index]