PcaGrid-class {rrcov}	R Documentation

Class "PcaGrid" - Robust PCA using PP - GRID search Algorithm

Description

Holds the results of an approximation of the PP-estimators for PCA using the grid search algorithm in the plane.

Objects from the Class

Objects can be created by calls of the form new("PcaGrid", ...) but the usual way of creating PcaGrid objects is a call to the function PcaGrid() which serves as a constructor.

Slots

call, center, scale, rank, loadings, eigenvalues, scores, k, sd, od, cutoff.sd, cutoff.od, flag, n.obs:

from the "Pca" class.

Extends

Class "PcaRobust", directly. Class "Pca", by class "PcaRobust", distance 2.

Methods

getQuan

signature(obj = "PcaGrid"): ...

Author(s)

Valentin Todorov valentin.todorov@chello.at

References

Todorov V & Filzmoser P (2009), An Object Oriented Framework for Robust Multivariate Analysis. Journal of Statistical Software, 32(3), 1–47. doi: 10.18637/jss.v032.i03.

See Also

PcaRobust-class, Pca-class, PcaClassic, PcaClassic-class

Examples

showClass("PcaGrid")

[Package rrcov version 1.7-7 Index]