PcaHubert-class {rrcov}	R Documentation

Class "PcaHubert" - ROBust method for Principal Components Analysis

Description

The ROBPCA algorithm was proposed by Hubert et al (2005) and stays for 'ROBust method for Principal Components Analysis'. It is resistant to outliers in the data. The robust loadings are computed using projection-pursuit techniques and the MCD method. Therefore ROBPCA can be applied to both low and high-dimensional data sets. In low dimensions, the MCD method is applied.

Objects from the Class

Objects can be created by calls of the form new("PcaHubert", ...) but the usual way of creating PcaHubert objects is a call to the function PcaHubert which serves as a constructor.

Slots

alpha:

Object of class "numeric" the fraction of outliers the algorithm should resist - this is the argument alpha

quan:

The quantile h used throughout the algorithm

skew:

Whether the adjusted outlyingness algorithm for skewed data was used

ao:

Object of class "Uvector" Adjusted outlyingness within the robust PCA subspace

call, center, scale, rank, loadings, eigenvalues, scores, k, sd, od, cutoff.sd, cutoff.od, flag, n.obs, eig0, totvar0:

from the "Pca" class.

Extends

Class "PcaRobust", directly. Class "Pca", by class "PcaRobust", distance 2.

Methods

getQuan

signature(obj = "PcaHubert"): Returns the quantile used throughout the algorithm

Author(s)

Valentin Todorov valentin.todorov@chello.at

References

Todorov V & Filzmoser P (2009), An Object Oriented Framework for Robust Multivariate Analysis. Journal of Statistical Software, 32(3), 1–47. doi: 10.18637/jss.v032.i03.

See Also

PcaRobust-class, Pca-class, PcaClassic, PcaClassic-class

Examples

showClass("PcaHubert")

[Package rrcov version 1.7-7 Index]