PredictQda-class {rrcov}	R Documentation

Class "PredictQda" - prediction of "Qda" objects

Description

The prediction of a "Qda" object

Objects from the Class

Objects can be created by calls of the form new("PredictQda", ...) but most often by invoking 'predict' on a "Qda" object. They contain values meant for printing by 'show'

Slots

classification:

a factor variable containing the classification of each object

posterior:

a matrix containing the posterior probabilities

x:

matrix with the discriminant scores

ct:

re-classification table of the training sample

Methods

show

signature(object = "PredictQda"): prints the results

Author(s)

Valentin Todorov valentin.todorov@chello.at

References

Todorov V & Filzmoser P (2009), An Object Oriented Framework for Robust Multivariate Analysis. Journal of Statistical Software, 32(3), 1–47. doi: 10.18637/jss.v032.i03.

See Also

Qda-class

Examples

showClass("PredictQda")

[Package rrcov version 1.7-7 Index]