QdaClassic {rrcov}	R Documentation

Quadratic Discriminant Analysis

Description

Performs quadratic discriminant analysis and returns the results as an object of class QdaClassic (aka constructor).

Usage

QdaClassic(x, ...)

## Default S3 method:
QdaClassic(x, grouping, prior = proportions, tol = 1.0e-4, ...)

Arguments

x
a matrix or data frame containing the explanatory variables (training set).

grouping
grouping variable: a factor specifying the class for each observation.

prior
prior probabilities, default to the class proportions for the training set.

tol
tolerance

...
arguments passed to or from other methods.

Value

Returns an S4 object of class QdaClassic

Author(s)

Valentin Todorov valentin.todorov@chello.at

References

Todorov V & Filzmoser P (2009), An Object Oriented Framework for Robust Multivariate Analysis. Journal of Statistical Software, 32(3), 1–47. doi: 10.18637/jss.v032.i03.

See Also

Qda-class, QdaClassic-class,

[Package rrcov version 1.7-7 Index]