QdaClassic-class {rrcov}	R Documentation

Class "QdaClassic" - Quadratic Discriminant Analysis

Description

Contains the results of classical Quadratic Discriminant Analysis

Objects from the Class

Objects can be created by calls of the form new("QdaClassic", ...) but the usual way of creating QdaClassic objects is a call to the function QdaClassic which serves as a constructor.

Slots

call:

The (matched) function call.

prior:

Prior probabilities used, default to group proportions

counts:

number of observations in each class

center:

the group means

cov:

the group covariance matrices

covinv:

the inverse of the group covariance matrices

covdet:

the determinants of the group covariance matrices

method:

a character string giving the estimation method used

X:

the training data set (same as the input parameter x of the constructor function)

grp:

grouping variable: a factor specifying the class for each observation.

control:

Object of class "CovControl" inherited from class Qda specifying which estimate and with what estimation options to use for the group means and covariances. It is always NULL for classical discriminant analysis.

Extends

Class "Qda", directly.

Methods

No methods defined with class "QdaClassic" in the signature.

Author(s)

Valentin Todorov valentin.todorov@chello.at

References

Todorov V & Filzmoser P (2009), An Object Oriented Framework for Robust Multivariate Analysis. Journal of Statistical Software, 32(3), 1–47. doi: 10.18637/jss.v032.i03.

See Also

QdaRobust-class, Qda-class, QdaClassic

Examples

showClass("QdaClassic")

[Package rrcov version 1.7-7 Index]