QdaCov {rrcov}	R Documentation

Robust Quadratic Discriminant Analysis

Description

Performs robust quadratic discriminant analysis and returns the results as an object of class QdaCov (aka constructor).

Usage

QdaCov(x, ...)

## Default S3 method:
QdaCov(x, grouping, prior = proportions, tol = 1.0e-4,
                 method = CovControlMcd(), ...)

Arguments

x
a matrix or data frame containing the explanatory variables (training set).

grouping
grouping variable: a factor specifying the class for each observation.

prior
prior probabilities, default to the class proportions for the training set.

tol
tolerance

method
method

...
arguments passed to or from other methods

Details

details

Value

Returns an S4 object of class QdaCov

Warning

Still an experimental version!

Author(s)

Valentin Todorov valentin.todorov@chello.at

References

Todorov V & Filzmoser P (2009), An Object Oriented Framework for Robust Multivariate Analysis. Journal of Statistical Software, 32(3), 1–47. doi: 10.18637/jss.v032.i03.

See Also

CovMcd

Examples

## Example anorexia
library(MASS)
data(anorexia)

## start with the classical estimates
qda <- QdaClassic(Treat~., data=anorexia)
predict(qda)@classification

## try now the robust LDA with the default method (MCD with pooled whitin cov matrix)
rqda <- QdaCov(Treat~., data= anorexia)
predict(rqda)@classification

## try the other methods
QdaCov(Treat~., data= anorexia, method="sde")
QdaCov(Treat~., data= anorexia, method="M")
QdaCov(Treat~., data= anorexia, method=CovControlOgk())

[Package rrcov version 1.7-7 Index]