biplot-methods {rrcov}	R Documentation

Biplot for Principal Components (objects of class 'Pca')

Description

Produces a biplot from an object (derived from) Pca-class.

Usage

    ## S4 method for signature 'Pca'
biplot(x, choices=1L:2L, scale=1, ...)

Arguments

x
an object of class (derived from) "Pca".

choices
length 2 vector specifying the components to plot. Only the default is a biplot in the strict sense.

scale
The variables are scaled by lambda ^ scale and the observations are scaled by lambda ^ (1-scale) where lambda are the singular values as computed by the Principal Components function. Normally 0 <= scale <= 1, and a warning will be issued if the specified scale is outside this range.

...
optional arguments to be passed to the internal graphical functions.

Side Effects

a plot is produced on the current graphics device.

Methods

biplot

signature(x = Pca): Plot a biplot, i.e. represent both the observations and variables of a matrix of multivariate data on the same plot. See also biplot.princomp.

References

Gabriel, K. R. (1971). The biplot graphical display of matrices with applications to principal component analysis. Biometrika, 58, 453–467.

See Also

Pca-class, PcaClassic, PcaRobust-class.

Examples

require(graphics)
biplot(PcaClassic(USArrests, k=2))

[Package rrcov version 1.7-7 Index]