restimate-methods {rrcov}	R Documentation

Methods for Function estimate in Package 'rrcov'

Description

Each concrete control class, like CovControlMest, CovControlOgk, etc., should implement an restimate method which will call the correponding (constructor)-function and will return the obtained S4 class, derived from CovRobust.

Usage

## S4 method for signature 'CovControlMest'
restimate(obj, x, ...)

Arguments

obj
an object of class "CovControlEstimate"

x
Data frame or matrix containing the data

.

...
other parameters to be passed through to the estimation function.

Methods

obj = "CovControlMcd"

Compute the MCD estimates of multivariate location and scatter by callingCovMcd

obj = "CovControlMest"

Compute the constrained M-estimates of multivariate location and scatter by callingCovMest

obj = "CovControlOgk"

Compute the Ortogonalized Gnanadesikan-Kettenring (OGK) estimates of multivariate location and scatter by callingCovOgk

[Package rrcov version 1.7-7 Index]