getCenter-methods {rrcov}	R Documentation

Accessor methods to the essential slots of Cov and its subclasses

Description

Accessor methods to the slots of objects of classCov and its subclasses

Usage

getCenter(obj)
getCov(obj)
getCorr(obj)
getData(obj)
getDistance(obj)
getEvals(obj)
getDet(obj)
getShape(obj)
getFlag(obj, prob=0.975)
getMeth(obj)
isClassic(obj)
getRaw(obj)

Arguments

obj
an object of class "Cov" or of a class derived from "Cov"

prob
optional argument for getFlag - probability, defaults to 0.975

Methods

obj = "Cov"

generic functions - see getCenter, getCov, getCorr, getData, getDistance, getEvals, getDet, getShape, getFlag, isClassic

obj = "CovRobust"

generic functions - see getCenter, getCov, getCorr, getData, getDistance, getEvals, getDet, getShape, getFlag, getMeth, isClassic

[Package rrcov version 1.7-7 Index]