getLoadings-methods {rrcov}	R Documentation

Accessor methods to the essential slots of Pca and its subclasses

Description

Accessor methods to the slots of objects of class Pca and its subclasses

Arguments

obj
an object of class "Pca" or of a class derived from "Pca"

Methods

obj = "Pca"

Accessors for object of class Pca

obj = "PcaRobust"

Accessors for object of class PcaRobust

obj = "PcaClassic"

Accessors for object of class PcaClassic

[Package rrcov version 1.7-7 Index]