pca.distances {rrcov}	R Documentation

Compute score and orthogonal distances for Principal Components (objects of class 'Pca')

Description

Compute score and orthogonal distances for an object (derived from)Pca-class.

Usage

    pca.distances(obj, data, r, crit=0.975)

Arguments

obj
an object of class (derived from) "Pca".

data
The data matrix for which the "Pca" object was computed.

r
rank of data

crit
Criterion to use for computing the cutoff values.

Details

This function calculates the score and orthogonal distances and the appropriate cutoff values for identifying outlying observations. The computed values are used to create a vector a of flags, one for each observation, identifying the outliers.

Value

An S4 object of class derived from the virtual class Pca-class - the same object passed to the function, but with the score and orthogonal distances as well as their cutoff values and the corresponding flags appended to it.

Author(s)

Valentin Todorov valentin.todorov@chello.at

References

M. Hubert, P. J. Rousseeuw, K. Vanden Branden (2005), ROBPCA: a new approach to robust principal components analysis, Technometrics, 47, 64–79.

Todorov V & Filzmoser P (2009), An Object Oriented Framework for Robust Multivariate Analysis. Journal of Statistical Software, 32(3), 1–47. doi: 10.18637/jss.v032.i03.

Examples

## PCA of the Hawkins Bradu Kass's Artificial Data
##  using all 4 variables
data(hbk)
pca <- PcaHubert(hbk)
pca.distances(pca, hbk, rankMM(hbk))

[Package rrcov version 1.7-7 Index]