pca.scoreplot {rrcov}	R Documentation

Score plot for Principal Components (objects of class 'Pca')

Description

Produces a score plot from an object (derived from) Pca-class.

Usage

    pca.scoreplot(obj, i=1, j=2, main, id.n, ...)

Arguments

obj
an object of class (derived from) "Pca".

i
First score coordinate, defaults to i=1.

j
Second score coordinate, defaults to j=2.

main
The main title of the plot.

id.n
Number of observations to identify by a label. If missing and the total number of observations is less or equal to 10, all observations will be labelled.

...
Optional arguments to be passed to the internal graphical functions.

Author(s)

Valentin Todorov valentin.todorov@chello.at

See Also

Pca-class, PcaClassic, PcaRobust-class.

Examples

require(graphics)

## PCA of the Hawkins Bradu Kass's Artificial Data
##  using all 4 variables
data(hbk)
pca <- PcaHubert(hbk)
pca
pca.scoreplot(pca)

[Package rrcov version 1.7-7 Index]