rrcov-utils {rrcov}	R Documentation

Different utility functions to be used in rrcov and packages depending on rrcov

Description

Different utility functions to be used in rrcov and packages depending on rrcov:

myscreeplot Plots screeplots for two covariance matrices

mtxconfusion Computes a confusion matrix between actual and predicted class membership

MATLAB-like functions

sqrtm Returns the square root of a symetric positive definite matrix

ones Returns a nxp matrix filled with ones

zeros Returns a nxp matrix filled with zeros

vecnorm Returns the norm of a vector

rankMM Returns the rank of the matrix

repmat Replicates a matrix

Usage

    myscreeplot(rcov, ccov)
    mtxconfusion(actual, predicted, prior = NULL, printit=FALSE)
    sqrtm(A)

Arguments

A
A matrix.

rcov, ccov
Covariance matrices.

actual
Actual class membership.

predicted
Predicted class membership.

prior
Prior probabilities.

printit
Wheather to print the results.

Author(s)

Valentin Todorov valentin.todorov@chello.at

[Package rrcov version 1.7-7 Index]