scorePlot-methods {rrcov}	R Documentation

Score plot for Principal Components (objects of class 'Pca')

Description

Produces a score plot from an object (derived from) Pca-class.

Usage

    ## S4 method for signature 'Pca'
scorePlot(x, i=1, j=2, ...)

Arguments

x
an object of class (derived from) "Pca".

i
First score coordinate, defaults to i=1.

j
Second score coordinate, defaults to j=2.

...
optional arguments to be passed to the internal graphical functions.

Side Effects

a plot is produced on the current graphics device.

Methods

scorePlot

signature(x = Pca): Plot a scatter plot of ith against jth score of the Pca object with superimposed tollerance (0.975) ellipse. See also biplot, screeplot.

See Also

Pca-class, PcaClassic, PcaRobust-class.

Examples

require(graphics)

## PCA of the Hawkins Bradu Kass's Artificial Data
##  using all 4 variables
data(hbk)
pca <- PcaHubert(hbk)
pca

scorePlot(pca)

[Package rrcov version 1.7-7 Index]