soil {rrcov}	R Documentation

Exchangable cations in forest soil data set

Description

The forest soil data set contains measurements on 58 soil pits in the Hubbard Brook Experimental Forest in north-central New Hampshire. The excavations were done in 1983 and 1986. The soil samples were analyzed for the exchangeable cations of aluminium, calcium, magnesium, potassium and sodium. The pit locations in both data sets can be classified by the type of the forest:

1: spruce-fir (11 samples),

2: high elevation hardwood (23 samples) and

3: low elevation hardwood (24 samples)).

Additionally the degree of logging disturbance can be considered (all 0 in the 1983 data set):

0: uncut forest,

1: cut, undisturbed by machinery and

2: cut, disturbed.

The observations are expressed in grams of exchangeable cations per square meter.

Usage

data(soil)

Format

A data frame with 116 observations on the following 7 variables.

F

Type of forest

D

Degree of logging disturbance

Al

Level of the exchangable cations in Al

Ca

Level of the exchangable cations in Ca

Mg

Level of the exchangable cations in Mg

K

Level of the exchangable cations in K

Na

Level of the exchangable cations in Na

Source

Morrison D.F., 2005, Multivariate Statistical Methods, Thompson

References

Vanden Branden K, Hubert M (2005). Robust Classiffication in High Dimensions Based on the SIMCA Method. Cbemometrics and Intelligent Laboratoty Sysiem, 79: 10–21.

Examples

data(soil)
soil1983 <- soil[soil$D == 0, -2]       # only 1983, remove column D (always 0)

(cc <- Linda(F~., data=soil))
(pr <- predict(cc))
pr@classification

[Package rrcov version 1.7-7 Index]