wolves {rrcov}	R Documentation

Skull dimensions of the wolf Canis lupus L.

Description

A data set containing skull morphometric measurements on Rocky Mountain and Arctic wolves (Canis Lupus L.). The tdata are published in Morrison (1990), originally from Jolicoeur (1959).

Usage

data(wolves)

Format

A data frame with 25 rows and 12 variables. The variables are as follows (all measurements are in milimeters):

class: a factor presenting the combinations of location and sex. The levels are arf arm rmf and rmm

location: a factor with levels ar=Arctic, rm=Rocky Mountain

sex: a factor with levels f=female, m=male

x1: palatal length

x2: postpalatal length

x3: zygomatic width

x4: palatal width outside first upper molars

x5: palatal width inside second upper molars

x6: postglenoid foramina width

x7: interorbital width

x8: braincase width

x9: crown length

Source

Jolicoeur, P. Multivariate geographical variation in the wolf Canis lupis L., Evolution, XIII, 283–299.

Morrison, D. F. Multivariate Statistical Methods, (3rd ed.), 1990. New York: McGraw-Hill, p. 288–289.

Examples

 data(wolves)

 ## Remove the factors location and sex which we will not use for now
 x <- wolves[,-c(2:3)]

 ## Plot a pairwise scaterplot matrix
 pairs(x[,2:10])

 mcd <- CovMcd(x[, 2:10])
 plot(mcd, which="pairs")

 lda <- LdaClassic(class~., data=x)
 lda@center
 lda@cov

 predict(lda)

[Package rrcov version 1.7-7 Index]